pyproject.toml
[tool.poetry]
name = "mailer"
version = "0.1.0"
description = ""
authors = ["II9tka <fffxx@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
environs = "^9.5.0"
Markdown = "^3.3.7"
django-celery-beat = "^2.2.1"
redis = "^4.3.1"
celery = "^5.2.6"
Django = "3.2"
django-fsm = "^2.8.0"
pytidylib = "^0.3.2"
pytracking = "^0.2.2"
django-solo = "^2.0.0"
lxml = "^4.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
